Documentation, remove test prints
