+++ src/irc_bridge/mod.rs
@@ -2,14 +2,16 @@ use std::sync::Arc;
 
 use bimap::BiMap;
 use irc::client::prelude::Config;
-use poise::serenity_prelude::Context;
+use poise::serenity_prelude::{Context, json};
 use tokio::sync::mpsc;
+
+use crate::Data;
 
 mod irc_half;
 mod discord_half;
 mod formatting;
 
-
+// A message crossing the bridge
 #[derive(Debug)]
 pub struct BridgeMessage {
     pub author: String,
@@ -17,9 +19,44 @@ pub struct BridgeMessage {
     pub message: String
 }
 
+// Channel for sending and receiving `BridgeMessage`s
 pub type Sender = mpsc::Sender<BridgeMessage>;
 pub type Receiver = mpsc::Receiver<BridgeMessage>;
 
+// Load user data with config options from the `irc` object in `config.json`
+pub fn load_data_from_config(data: &mut Data, irc_config: &json::Value) {
+    // Mark that the IRC bridge has not started yet
+    data.irc_running_bridge = false.into();
+    // Load channels into the channel map and the IRC config object
+    let channels = irc_config["channels"]
+        .as_object()
+        .expect("config.json: `irc.channels` must exist if `irc` exists");
+    for (k, v) in channels {
+        let irc_channel = k.to_owned();
+        let dc_channel = v.as_u64().expect("config.json: values in `irc.channels` must be integers");
+        data.irc_channel_map.insert(dc_channel, irc_channel.clone());
+        data.irc_config.channels.push(irc_channel);
+    }
+    // set the IRC server
+    data.irc_config.server = Some(irc_config["server"]
+        .as_str()
+        .expect("config.json: `irc.server` must exist if `irc` exists")
+        .to_owned()
+    );
+    // set the IRC channels
+    data.irc_config.nickname = Some(irc_config["nickname"]
+        .as_str()
+        .expect("config.json: `irc.nickname` must exist if `irc` exists")
+        .to_owned()
+    );
+    // enable/disable TLS (default: enabled)
+    data.irc_config.use_tls = irc_config["use_tls"].as_bool();
+}
+
+// Start running the IRC bridge. This function creates two new
+// asynchronous tasks, one for the IRC client and one that
+// transmits new messages to Discord. Recieving messages is
+// done by the main program via the returned `Sender`.
 pub fn run(
     ctx: Arc<Context>, 
     irc_config: Config, 
@@ -36,7 +73,7 @@ pub fn run(
     });
     tokio::spawn(async {
         if let Err(e) = discord_half::run_bridge(ctx, rx_id, channel_mapping).await {
-            println!("Error in IRC bridge: {}", e);
+            println!("Error in Discord bridge: {}", e);
         }
     });
     return tx_di;
